refactor(api): déplacer interactors et resolver dans ::API
DataPackages::Transmit et DeliveryCriteriaResolver n'appartiennent
qu'à l'API gelée — ils n'ont aucune raison de vivre dans :: global.

- app/interactors/data_packages/ → app/interactors/api/data_packages/
- app/queries/delivery_criteria_resolver → app/queries/api/
- Références mises à jour dans les controllers concernés
