+++ app/interactors/api/data_packages/transmit.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module API
+  module DataPackages
+    class Transmit
+      include Interactor::Organizer
+
+      organize Transmit::ValidateTransmission,
+        Transmit::ResolveRecipients,
+        Transmit::CreateNotifications,
+        Transmit::TransitionToTransmitted
+    end
+  end
+end
