--- app/interactors/data_packages/transmit.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-module DataPackages
-  class Transmit
-    include Interactor::Organizer
-
-    organize Transmit::ValidateTransmission,
-      Transmit::ResolveRecipients,
-      Transmit::CreateNotifications,
-      Transmit::TransitionToTransmitted
-  end
-end
